+++ gulpfile.js
@@ -13,6 +13,16 @@ require('laravel-elixir-vue');
  |
  */
 
+// Include JS from bower
+jsFromBower = (scripts) => {
+    bower_scripts = []
+    scripts.forEach((script) => {
+        bower_scripts.push(`../bower/${script}.js`)
+    })
+    return bower_scripts
+}
+
+
 elixir(mix => {
     mix
         .browserSync({
@@ -25,30 +35,30 @@ elixir(mix => {
         .sass('app.scss')
         .coffee(['resources/assets/coffee/*.coffee', 'resources/assets/coffee/*/*.coffee'])
         .copy('resources/assets/bower/SVG-Loaders/svg-loaders/three-dots.svg', 'public/img/svg')
-        .scripts([
-            '../bower/jquery/dist/jquery.js',
-            '../bower/bootstrap/dist/js/bootstrap.min.js',
-            '../bower/angular/angular.min.js',
-            '../bower/angular-animate/angular-animate.min.js',
-            '../bower/angular-sanitize/angular-sanitize.min.js',
-            '../bower/angular-resource/angular-resource.min.js',
-            '../bower/angular-aria/angular-aria.min.js',
-            '../bower/angular-messages/angular-messages.min.js',
-            '../bower/angular-material/angular-material.min.js',
-            '../bower/angular-i18n/angular-locale_ru-ru.js',
-            '../bower/angular-inview/angular-inview.js',
-            '../bower/nprogress/nprogress.js',
-            '../bower/underscore/underscore-min.js',
-            '../bower/bootstrap-select/dist/js/bootstrap-select.js',
-            '../bower/bootstrap-datepicker/dist/js/bootstrap-datepicker.js',
-            '../bower/bootstrap-datepicker/dist/locales/bootstrap-datepicker.ru.min.js',
-            '../bower/angular-ui-sortable/sortable.min.js',
-            '../bower/angular-bootstrap/ui-bootstrap.min.js',
-            '../bower/ladda/dist/ladda.min.js',
-            '../bower/angular-ladda/dist/angular-ladda.min.js',
-            '../bower/moment/min/moment.min.js',
-            '../bower/moment/locale/ru.js',
-            '../bower/jquery.maskedinput/dist/jquery.maskedinput.min.js',
-            'resources/assets/js/*.js',
-        ], 'public/js/vendor.js');
+        .scripts(jsFromBower([
+            'jquery/dist/jquery',
+            'bootstrap/dist/js/bootstrap.min',
+            'angular/angular.min',
+            'angular-animate/angular-animate.min',
+            'angular-sanitize/angular-sanitize.min',
+            'angular-resource/angular-resource.min',
+            'angular-aria/angular-aria.min',
+            'angular-messages/angular-messages.min',
+            'angular-material/angular-material.min',
+            'angular-i18n/angular-locale_ru-ru',
+            'angular-inview/angular-inview',
+            'nprogress/nprogress',
+            'underscore/underscore-min',
+            'bootstrap-select/dist/js/bootstrap-select',
+            'bootstrap-datepicker/dist/js/bootstrap-datepicker',
+            'bootstrap-datepicker/dist/locales/bootstrap-datepicker.ru.min',
+            'angular-ui-sortable/sortable.min',
+            'angular-bootstrap/ui-bootstrap.min',
+            'ladda/dist/spin.min',
+            'ladda/dist/ladda.min',
+            'angular-ladda/dist/angular-ladda.min',
+            'moment/min/moment.min',
+            'moment/locale/ru',
+            'jquery.maskedinput/dist/jquery.maskedinput.min'
+        ]).concat(['resources/assets/js/*.js']), 'public/js/vendor.js');
 });